+++ examples/simple/foo/views.py
@@ -81,10 +81,55 @@ def authenticate(request, template_name='foo/authenticate.html'):
             (uid, remote_ska_login_url, signed_remote_ska_login_url)
         )
 
+    # Fail-only login URLs by provider
+    fail_only_remote_ska_login_urls_by_provider = []
+    uid = 'client_1.admins'
+    data = PROVIDERS[uid]
+    signed_remote_ska_login_url_invalid_username = sign_url(
+        auth_user='forbidden_username',
+        # Using provider-specific secret key
+        secret_key=data.get('SECRET_KEY'),
+        url=remote_ska_login_url,
+        extra={
+            'email': 'test_ska_user_{0}@mail.example.com'.format(uid),
+            'first_name': 'John {0}'.format(uid),
+            'last_name': 'Doe {0}'.format(uid),
+            DEFAULT_PROVIDER_PARAM: uid,
+        }
+    )
+    signed_remote_ska_login_url_invalid_email = sign_url(
+        auth_user='test_ska_user_{0}'.format(uid),
+        # Using provider-specific secret key
+        secret_key=data.get('SECRET_KEY'),
+        url=remote_ska_login_url,
+        extra={
+            'email': 'forbidden@example.com',
+            'first_name': 'John {0}'.format(uid),
+            'last_name': 'Doe {0}'.format(uid),
+            DEFAULT_PROVIDER_PARAM: uid,
+        }
+    )
+    fail_only_remote_ska_login_urls_by_provider.append(
+        (
+            uid,
+            'invalid username: forbidden_username',
+            signed_remote_ska_login_url_invalid_username,
+        )
+    )
+    fail_only_remote_ska_login_urls_by_provider.append(
+        (
+            uid,
+            'invalid email: forbidden@example.com',
+            signed_remote_ska_login_url_invalid_email
+        )
+    )
+
     # Template context
     context = {
         'remote_ska_login_urls': remote_ska_login_urls,
-        'remote_ska_login_urls_by_provider': remote_ska_login_urls_by_provider
+        'remote_ska_login_urls_by_provider': remote_ska_login_urls_by_provider,
+        'fail_only_remote_ska_login_urls_by_provider':
+            fail_only_remote_ska_login_urls_by_provider,
     }
 
     return render(request, template_name, context)