+++ examples/simple/settings/base.py
@@ -288,6 +288,7 @@ SKA_PROVIDERS = {
     # Client 1, group admins
     'client_1.admins': {
         'SECRET_KEY': 'client-1-admins-secret-key',
+        'USER_VALIDATE_CALLBACK': 'foo.ska_callbacks.client1_admins_validate',
         'USER_CREATE_CALLBACK': 'foo.ska_callbacks.client1_admins_create',
         'USER_GET_CALLBACK': 'foo.ska_callbacks.client1_admins_get',
         'USER_INFO_CALLBACK': 'foo.ska_callbacks.client1_admins_info',